test: add and merge user lifecycle tests into user-lifecyle
